adcom1: review numeric types - use fixed-size int/uint

+++ adcom1/content.go
@@ -20,7 +20,7 @@ type Content struct {
 	//   integer
 	// Definition:
 	//   Episode number.
-	Episode int `json:"episode,omitempty"`
+	Episode uint16 `json:"episode,omitempty"`
 
 	// Attribute:
 	//   title
